feat: require user authentication for all todo endpoints

+++ LM/project-3/TodoApp/routers/todos.py
@@ -6,6 +6,8 @@ from models import Todos
 from pydantic import BaseModel, Field
 from sqlalchemy.orm import Session
 
+from .auth import get_current_user
+
 router = APIRouter()
 
 
@@ -18,6 +20,7 @@ def get_db():
 
 
 db_depenency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
 class TodoRequest(BaseModel):
@@ -28,28 +31,39 @@ class TodoRequest(BaseModel):
 
 
 @router.get("/")
-async def read_all(db: db_depenency):
-    return db.query(Todos).all()
+async def read_all(db: db_depenency, user: user_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    return db.query(Todos).filter(Todos.owner_id == user.get("id")).all()
 
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(todo_id: Annotated[int, Path(..., gt=0)], db: db_depenency):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(todo_id: Annotated[int, Path(..., gt=0)], db: db_depenency, user: user_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail="Todo not found")
 
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(todo: TodoRequest, db: db_depenency):
-    todo_model = Todos(**todo.model_dump())
+async def create_todo(todo: TodoRequest, db: db_depenency, user: user_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    todo_model = Todos(**todo.model_dump(), owner_id=user.get("id"))
     db.add(todo_model)
     db.commit()
 
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(todo: TodoRequest, db: db_depenency, todo_id: Annotated[int, Path(..., gt=0)]):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(
+    todo: TodoRequest, db: db_depenency, user: user_dependency, todo_id: Annotated[int, Path(..., gt=0)]
+):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
 
@@ -62,8 +76,10 @@ async def update_todo(todo: TodoRequest, db: db_depenency, todo_id: Annotated[in
 
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(todo_id: Annotated[int, Path(..., gt=0)], db: db_depenency):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(todo_id: Annotated[int, Path(..., gt=0)], db: db_depenency, user: user_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
     db.delete(todo_model)
